requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-schema-viewer==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
kombu==5.3.4
prompt-toolkit==3.0.41
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
